Refresh requirements (#576)
* Require Numba 0.57+ consistently
* Require NumPy 1.21+ (to align with [Numba 0.57.0's NumPy minimum]( https://github.com/numba/numba/blob/0.57.0/setup.py#L25 ))
* Allow NumPy 1.24 (soften upper bounds)
* Replace deprecated `np.int` with `int`

Authors:
  - https://github.com/jakirkham

Approvers:
  - AJ Schmidt (https://github.com/ajschmidt8)
  - Adam Thompson (https://github.com/awthomp)
  - Bradley Dice (https://github.com/bdice)
  - Ray Douglass (https://github.com/raydouglass)

URL: https://github.com/rapidsai/cusignal/pull/576

+++ python/cusignal/convolution/_convolution_cuda.py
@@ -400,7 +400,7 @@ def _convolve(
     in2 = in2.astype(promType)
 
     # Create empty array to hold number of aout dimensions
-    out_dimens = np.empty(in1.ndim, np.int)
+    out_dimens = np.empty(in1.ndim, int)
     if val == VALID:
         for i in range(in1.ndim):
             out_dimens[i] = (
@@ -465,7 +465,7 @@ def _convolve2d(in1, in2, use_convolve, mode, boundary, fillvalue):
             raise Exception("Unable to create fill array")
 
     # Create empty array to hold number of aout dimensions
-    out_dimens = np.empty(in1.ndim, np.int)
+    out_dimens = np.empty(in1.ndim, int)
     if val == VALID:
         for i in range(in1.ndim):
             out_dimens[i] = in1.shape[i] - in2.shape[i] + 1
@@ -539,7 +539,7 @@ def _convolve1d2o(in1, in2, mode):
     in2 = in2.astype(promType)
 
     # Create empty array to hold number of aout dimensions
-    out_dimens = np.empty(in1.ndim, np.int)
+    out_dimens = np.empty(in1.ndim, int)
     if val == VALID:
         for i in range(in1.ndim):
             out_dimens[i] = in1.shape[i] - in2.shape[i] + 1
@@ -602,7 +602,7 @@ def _convolve1d3o(in1, in2, mode):
     in2 = in2.astype(promType)
 
     # Create empty array to hold number of aout dimensions
-    out_dimens = np.empty(in1.ndim, np.int)
+    out_dimens = np.empty(in1.ndim, int)
     if val == VALID:
         for i in range(in1.ndim):
             out_dimens[i] = in1.shape[i] - in2.shape[i] + 1
